pyproject.toml
[tool.poetry]
name = "gop-report-generator"
version = "0.1.0"
description = ""
authors = ["saogregl <lsangregorio2@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
pandas = "^2.2.0"
flatdict = "4.0.1"
XlsxWriter = "^3.1.9"
openpyxl = "^3.1.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
